Reprocess public_image originals in the worker; drop dead dimension guard
Fallout from backgrounding public_image version generation:

- Backgrounder::Delay skips the original's top-level processors on upload,
  and the worker only recreated versions, so fix_exif_rotation and strip
  stopped running. Originals (served as the just-uploaded fallback and the
  full-size link) kept their EXIF orientation and metadata, incl. GPS. The
  worker now re-caches and re-stores the original, rerunning those processors
  and rebuilding every version from the processed original in one pass.
- validate_dimensions was dead in the deferred path and would instead fail
  noisily deep in the worker once the original is reprocessed, so drop it.
  Avatar/ManufacturerLogo keep their own synchronous guard.
- Consolidate the near-duplicate PublicImageProcessJob into
  CarrierWaveProcessJob, which does the unscoped lookup for both public
  images and Bike#pdf.
