+++ spec/jobs/images/external_url_store_job_spec.rb
@@ -23,7 +23,7 @@ RSpec.describe Images::ExternalUrlStoreJob, type: :job do
       # AfterBikeSaveJob + backgrounded version generation (process_in_background :image)
       expect(Sidekiq::Job.jobs.count).to eq 2
       expect(CallbackJob::AfterBikeSaveJob).to have_enqueued_sidekiq_job(bike.id, false, true)
-      expect(PublicImageProcessJob).to have_enqueued_sidekiq_job("PublicImage", public_image.id.to_s, "image")
+      expect(CarrierWaveProcessJob).to have_enqueued_sidekiq_job("PublicImage", public_image.id.to_s, "image")
     end
     context "is_private true" do
       let(:is_private) { true }
@@ -39,7 +39,7 @@ RSpec.describe Images::ExternalUrlStoreJob, type: :job do
         # AfterBikeSaveJob + backgrounded version generation (process_in_background :image)
         expect(Sidekiq::Job.jobs.count).to eq 2
         expect(CallbackJob::AfterBikeSaveJob).to have_enqueued_sidekiq_job(bike.id, false, true)
-        expect(PublicImageProcessJob).to have_enqueued_sidekiq_job("PublicImage", public_image.id.to_s, "image")
+        expect(CarrierWaveProcessJob).to have_enqueued_sidekiq_job("PublicImage", public_image.id.to_s, "image")
       end
     end
   end
